Create Reverse_Asc_File
Sometimes reading a file/textstream backwards (from end tot start) is required. For example, the central directory structue of an NTFS compressed file (pkzip) is found at the end of the file. Reading backwards a textstream  is not possible without extra manipulation. This procedure makes a copy of the sourcefile in the reversed order.